editing

+++ backend/src/main/java/com/tfg/backend/repository/PropertyRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface PropertyRepository extends JpaRepository<Property, Long> {
+public interface PropertyRepository extends JpaRepository<Property, Integer> {
 
     List<Property> findByTypeOrderByPublishDateDesc(Property.PropertyType type);
